feat:add delete route tu delete file into MongoDB as well as localhost

+++ backend/controllers/stuff.js
@@ -1,5 +1,7 @@
 const Thing = require('../models/Thing.js')
 
+const fs = require('fs')
+
 exports.createThing = (req, res, next) => {
     const thingObject = JSON.parse(req.body.thing);
     delete thingObject._id;
@@ -28,7 +30,7 @@ exports.modifyThing = (req, res, next) => {
                 res.status(401).json({ message: 'Non authorisé' });
             } else {
                 Thing.updateOne({ _id: req.params.id }, { ...thingObject, _id: req.params.id })
-                    .then(thing => res.status(200).json({ message: 'Objet modifié' }))
+                    .then(() => res.status(200).json({ message: 'Objet modifié' }))
                     .catch(error => res.status(401).json({ error }))
             }
         })
@@ -39,9 +41,24 @@ exports.modifyThing = (req, res, next) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé' }))
-        .catch(error => res.status(400).json({ error }));
+    Thing.findOne({ _id: req.params.id })
+        .then(thing => {
+            if (thing.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Non authorisé' });
+            } else {
+                const filename = thing.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Thing.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: 'Objet supprimé' }))
+                        .catch(error => res.status(400).json({ error }));
+                })
+            }
+        })
+        .catch((error) => {
+            res.status(500).json({ error });
+        })
+
+
 }
 
 exports.getOneThing = (req, res, next) => {
